Font installation is now conditional

+++ scripts/ubuntu_configure_zsh.sh
@@ -5,6 +5,7 @@ set -eu
 # variables
 font=IosevkaTerm
 nerdfonts_version=v3.2.1
+font_dir=~/.local/share/fonts
 
 # make zsh the default shell
 echo "Setting zsh as the default shell"
@@ -18,13 +19,15 @@ else
 fi
 
 # download the Iosevka font
-echo "Downloading and installing the IosevkaTerm font"
-mkdir -p ~/.local/share/fonts
-wget -O ~/.local/share/fonts/$font.zip https://github.com/ryanoasis/nerd-fonts/releases/download/$nerdfonts_version/$font.zip
-cd ~/.local/share/fonts
-unzip -o $font.zip
-rm $font.zip
-fc-cache -fv
+if [ ! -f $font_dir/"$font"NerdFont-Regular.ttf ]; then
+	echo "Downloading and installing the IosevkaTerm font"
+	mkdir -p $font_dir
+	wget -O $font_dir/$font.zip https://github.com/ryanoasis/nerd-fonts/releases/download/$nerdfonts_version/$font.zip
+	cd $font_dir
+	unzip -o $font.zip
+	rm $font.zip
+	fc-cache -fv
+fi
 
 # install the Iosevka font as the terminal default
 dconf write /org/gnome/terminal/legacy/profiles:/:"$(gsettings get org.gnome.Terminal.ProfilesList default | tr -d "'")"/font "'$font Nerd Font 12'"
